% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQpaste.R
\name{EQpaste}
\alias{EQpaste}
\title{The function creates a variable, STATE, that is all the values together and can be used with the QALY function}
\usage{
EQpaste(x)
}
\arguments{
\item{EQpaste}{Converts the columns into STATE which is the pasted values}
}
\value{
STATE which is used im the QALY function
}
\description{
This only works if the data conforms to the standard naming of EQ-5D-5L variables according to the user 
guide Which can be found at: https://euroqol.org/wp-content/uploads/2016/09/EQ-5D-5L_UserGuide_2015.pdf
}
\details{
The naming of the data should be: MOBILITY, SELFCARE, ACTIVITY, PAIN, ANXIETY
}
\examples{
EQpaste(df)
}
