% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QALY.R
\name{QALY}
\alias{QALY}
\title{Calculates EQ-5D-5L - Quality Adjusted Life Years}
\usage{
QALY(x, country = "dk", name = "STATE")
}
\arguments{
\item{QALY}{Converts the questionare scores}
}
\value{
Quality Adjusted Life Years that ranges between 1, and slightly negative, meaning worse than dead
}
\description{
This function converts scores from the EQ-5D-5L questionaire to Quality Adjusted Life Years. 
The function do this by using value sets of different countries:
}
\details{
Denmark	France	Germany	Japan	Netherlands	Spain	Thailand	UK	US	Zimbabwe

To use it scores of the 5 dimensions need to be pasted together and stored in a dataframe column.
eg. 11111 is someone with good health as indicated by answers in the 5 categories and corronsponds
to a QALY of 1, except in Zimbabwe, while 53454 would be someone with a bad health and would corronspond 
to a low or negative QALY
}
\examples{
QALY(df, country="dk", name="pasted")
}
